+++ nix/cells/automation/ciJobs.nix
@@ -63,15 +63,12 @@ let
         ghc810.roots = x86_64-linux.ghc810.roots;
         ghc92.roots = x86_64-linux.ghc92.roots;
         ghc96.roots = x86_64-linux.ghc96.roots;
-        packages = x86_64-linux.packages;
         # Note: We can't build the 9.6 shell on aarch64-darwin
         # because of https://github.com/well-typed/cborg/issues/311
         devshells = removeAttrs x86_64-linux.devshells [ "plutus-shell-96" ];
       };
 
-      aarch64-linux = { };
-
-      system-matrix = { inherit x86_64-linux x86_64-darwin aarch64-darwin aarch64-linux; };
+      system-matrix = { inherit x86_64-linux x86_64-darwin aarch64-darwin; };
     in
     system-matrix.${system};
 
@@ -85,7 +82,11 @@ let
     constituents = lib.collect lib.isDerivation filtered-jobs;
   };
 
-  final-jobset = filtered-jobs // { required = required-job; };
+  merged-jobset = filtered-jobs // { required = required-job; };
+
+  is-supported-system = lib.elem system [ "x86_64-linux" "x86_64-darwin" "aarch64-darwin" ];
+
+  final-jobset = lib.optionalAttrs is-supported-system merged-jobset;
 
 in