+++ plutus-core/plutus-core/src/PlutusCore/Default/Universe.hs
@@ -208,7 +208,7 @@ instance HasUniApply DefaultUni where
 deriving stock instance Show (DefaultUni a)
 instance GShow DefaultUni where gshowsPrec = showsPrec
 
-instance HasRenderContext config => PrettyBy config (DefaultUni a) where
+instance PrettyBy RenderContext (DefaultUni a) where
     prettyBy = inContextM $ \case
         DefaultUniInteger              -> "integer"
         DefaultUniByteString           -> "bytestring"
@@ -223,13 +223,13 @@ instance HasRenderContext config => PrettyBy config (DefaultUni a) where
         DefaultUniBLS12_381_G2_Element -> "bls12_381_G2_element"
         DefaultUniBLS12_381_MlResult   -> "bls12_381_mlresult"
 
-instance HasRenderContext config => PrettyBy config (SomeTypeIn DefaultUni) where
+instance PrettyBy RenderContext (SomeTypeIn DefaultUni) where
     prettyBy config (SomeTypeIn uni) = prettyBy config uni
 
 -- | This always pretty-prints parens around type applications (e.g. @(list bool)@) and
 -- doesn't pretty-print them otherwise (e.g. @integer@).
 instance Pretty (DefaultUni a) where
-    pretty = prettyParens
+    pretty = prettyBy juxtRenderContext
 instance Pretty (SomeTypeIn DefaultUni) where
     pretty (SomeTypeIn uni) = pretty uni
 
@@ -346,9 +346,6 @@ So, what to do? We adopt the following strategy:
 
 Doing this effectively bans builds on 32-bit systems, but that's fine, since we don't care about
 supporting 32-bit systems anyway, and this way any attempts to build on them will fail fast.
-
-Note: We have another 64-bit limitation, this time not during script execution but during
-script deserialization, for more see Note [Index (Word64) (de)serialized through Natural].
 
 Note: we couldn't fail the bounds check with 'AsUnliftingError', because an out-of-bounds error
 is not an internal one -- it's a normal evaluation failure, but unlifting errors