+++ plutus-tx-plugin/test/Plugin/Spec.hs
@@ -10,6 +10,7 @@ import Plugin.Errors.Spec
 import Plugin.Functions.Spec
 import Plugin.Laziness.Spec
 import Plugin.NoTrace.Spec
+import Plugin.Optimization.Spec
 import Plugin.Primitives.Spec
 import Plugin.Profiling.Spec
 import Plugin.Strict.Spec
@@ -24,6 +25,7 @@ tests = testNested "Plugin" [
   , functions
   , laziness
   , noTrace
+  , optimization
   , errors
   , typeclasses
   , strict