+++ plutus-tx/src/PlutusTx/Builtins/Internal.hs
@@ -586,6 +586,14 @@ bls12_381_G1_hashToGroup (BuiltinByteString msg) (BuiltinByteString dst) =
       Left err -> mustBeReplaced $ show err
       Right p  -> BuiltinBLS12_381_G1_Element p
 
+{-# NOINLINE bls12_381_G1_zero #-}
+bls12_381_G1_zero ::  BuiltinBLS12_381_G1_Element
+bls12_381_G1_zero = BuiltinBLS12_381_G1_Element BLS12_381.G1.zero
+
+{-# NOINLINE bls12_381_G1_generator #-}
+bls12_381_G1_generator ::  BuiltinBLS12_381_G1_Element
+bls12_381_G1_generator = BuiltinBLS12_381_G1_Element BLS12_381.G1.generator
+
 ---------------- G2 ----------------
 
 data BuiltinBLS12_381_G2_Element = BuiltinBLS12_381_G2_Element BLS12_381.G2.Element
@@ -632,6 +640,14 @@ bls12_381_G2_hashToGroup (BuiltinByteString msg) (BuiltinByteString dst) =
     case BLS12_381.G2.hashToGroup msg dst of
       Left err -> mustBeReplaced $ show err
       Right p  -> BuiltinBLS12_381_G2_Element p
+
+{-# NOINLINE bls12_381_G2_zero #-}
+bls12_381_G2_zero ::  BuiltinBLS12_381_G2_Element
+bls12_381_G2_zero = BuiltinBLS12_381_G2_Element BLS12_381.G2.zero
+
+{-# NOINLINE bls12_381_G2_generator #-}
+bls12_381_G2_generator ::  BuiltinBLS12_381_G2_Element
+bls12_381_G2_generator = BuiltinBLS12_381_G2_Element BLS12_381.G2.generator
 
 ---------------- Pairing ----------------